misc/ishihara/src/requirements.txt
argparse
pillow
scipy
numpy
colorlog
tqdm